[Refactor] Update module names
Recent extensions to version 1 of the Nest API have diverged from the
JSONAPI draft specification. Rename modules to avoid confusion between
the latest draft specification and the current reality of the responses.
